Add q10 elements to nml2 channels. Extend test coverage.
Now SampleCellNML2's K and Na channels are both specified in NeuroMLv2
and their rates have a q10 element with experimentalTemp=32degC, which
is the temperature the relevant simulation is run at. A variant of the
cell (SampleCellNML2_lowQ10ExpTemp) with experimentalTemp=25deg has
been added, and its spiketimes have been measured and included in the
relevant test script.